feat: Extract API token from request headers or basic auth
This commit modifies the `ProxyRequest` function in `proxy.go` to extract the API token from either the request headers or the basic authentication. If the request has basic authentication, the username is used as the API token. Otherwise, the API token is retrieved from the `X-Clickhouse-User` header. The commit also removes the `X-Clickhouse-User` and `X-Clickhouse-Key` headers from the request.